requirements.txt
flask
openai==0.27.6
python-dotenv==1.0.0